requirements-mac.txt
# Office OS — Mac-local requirements (Apple Silicon M1/M2/M3/M4)
# Inference via Ollama + optional LoRA training via MLX

# === Core simulation ===
gymnasium>=0.29.0
numpy>=1.24.0
pydantic>=2.0

# === LLM inference (Ollama uses OpenAI-compatible API) ===
openai>=1.0

# === Demo API server ===
fastapi>=0.115
uvicorn[standard]>=0.30
websockets>=12
python-dotenv>=1.0
httpx>=0.27
python-multipart>=0.0.9

# === Optional: Anthropic API (for Claude-based runs or benchmarking) ===
anthropic>=0.40

# === Optional: AWS Bedrock (if using Bedrock provider) ===
# boto3>=1.34

# === Optional: MLX for local LoRA training on Apple Silicon ===
# Uncomment these to enable local training:
# mlx>=0.22.0
# mlx-lm>=0.21.0

# === Optional: Frontend dev ===
# Node.js 18+ required for frontend (install via brew: brew install node)
